requirements.txt
numpy
pandas
matplotlib
PyQt5
Pillow
pytesseract
FuzzyWuzzy
QtRangeSlider
opencv
